+++ src/FileSystem.cpp
@@ -21,15 +21,23 @@ void FileSystem::load(RED4ext::PluginHandle p_handle,
   auto path = std::filesystem::absolute(".");
 
   game_path = path.parent_path().parent_path();
-  storages_path =
-    game_path / "red4ext" / "plugins" / "RedFileSystem" / "storages";
+  storages_path = game_path / "r6" / "storages";
   bool is_present = request_directory(storages_path);
 
   if (!is_present) {
     has_error = true;
     logger->ErrorF(handle, "Failed to create directory at \"%s\".",
-                   storages_path.c_str());
+                   storages_path.string().c_str());
     logger->Error(handle, "RedFileSystem has been disabled.");
+    return;
+  }
+  auto old_path =
+    game_path / "red4ext" / "plugins" / "RedFileSystem" / "storages";
+
+  if (!migrate_directory(old_path, storages_path)) {
+    logger->WarnF(handle, R"(Failed to migrate directory from "%s" to "%s".)",
+                  old_path.string().c_str(), storages_path.string().c_str());
+    logger->Warn(handle, "You need to manually move content yourself.");
     return;
   }
   has_error = false;
@@ -37,6 +45,11 @@ void FileSystem::load(RED4ext::PluginHandle p_handle,
 }
 
 void FileSystem::unload() {
+  std::error_code error;
+  auto old_path =
+    game_path / "red4ext" / "plugins" / "RedFileSystem" / "storages";
+
+  std::filesystem::remove_all(old_path, error);
   storages.clear();
   has_error = true;
   logger->Info(handle, "RedFileSystem has been terminated.");
@@ -53,7 +66,8 @@ inline Red::Handle<FileSystemStorage> FileSystem::get_storage(
   std::string name = p_name.c_str();
 
   if (!regex_match(name, storage_name_rule)) {
-    logger->ErrorF(handle, "Name of storage \"%s\" is not allowed.", name.c_str());
+    logger->ErrorF(handle, "Name of storage \"%s\" is not allowed.",
+                   name.c_str());
     logger->Error(handle, "See the documentation to fix this issue.");
     return {};
   }
@@ -73,8 +87,7 @@ inline Red::Handle<FileSystemStorage> FileSystem::get_storage(
   bool is_present = request_directory(path);
 
   if (!is_present) {
-    logger->ErrorF(handle, "Failed to create storage \"%s\".",
-                   name.c_str());
+    logger->ErrorF(handle, "Failed to create storage \"%s\".", name.c_str());
     return {};
   }
   auto storage = Red::MakeHandle<FileSystemStorage>(name, path);
@@ -99,6 +112,32 @@ bool FileSystem::request_directory(const std::filesystem::path& p_path) {
   if (!is_present || error) {
     return false;
   }
+  return true;
+}
+
+bool FileSystem::migrate_directory(const std::filesystem::path& p_old_path,
+                                   const std::filesystem::path& p_new_path) {
+  std::error_code error;
+  bool has_old_path = std::filesystem::exists(p_old_path, error);
+
+  if (error) {
+    return false;
+  }
+  if (!has_old_path) {
+    return true;
+  }
+  std::filesystem::copy_options options =
+    std::filesystem::copy_options::overwrite_existing |
+    std::filesystem::copy_options::recursive;
+
+  std::filesystem::copy(p_old_path, p_new_path, options, error);
+  if (error) {
+    logger->ErrorF(handle, "Could not migrate \"storages\" due to: %s.",
+                   error.message().c_str());
+    return false;
+  }
+  // NOTE: remove old path in unload() to be backward compatible with
+  //       RED4ext v1.24.3
   return true;
 }